# [chore] Speedup IsKnownFalsePositive using sets